+++ packages/ui/Scroller/Scroller.js
@@ -4,6 +4,7 @@
  * @module ui/Scroller
  * @exports Scroller
  * @exports ScrollerBase
+ * @exports ScrollerBasic
  * @exports ScrollerNative
  */
 
@@ -16,7 +17,7 @@ import {ResizeContext} from '../Resizable';
 import useScroll from '../useScroll';
 import Scrollbar from '../useScroll/Scrollbar';
 
-import {UiScrollerBase} from './UiScrollerBase';
+import {ScrollerBase} from './UiScrollerBase';
 
 import css from './Scroller.module.less';
 
@@ -25,15 +26,15 @@ import css from './Scroller.module.less';
  *
  * In most circumstances, you will want to use the Scrollable version.
  *
- * @class ScrollerBase
+ * @class ScrollerBasic
  * @memberof ui/Scroller
  * @ui
  * @public
  */
-class ScrollerBase extends Component {
-	static displayName = 'ui:ScrollerBase'
+class ScrollerBasic extends Component {
+	static displayName = 'ui:ScrollerBasic'
 
-	static propTypes = /** @lends ui/Scroller.ScrollerBase.prototype */ {
+	static propTypes = /** @lends ui/Scroller.ScrollerBasic.prototype */ {
 		children: PropTypes.node.isRequired,
 
 		/**
@@ -247,7 +248,7 @@ class ScrollerBase extends Component {
  * ```
  *
  * @name cbScrollTo
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {Function}
  * @public
  */
@@ -261,7 +262,7 @@ class ScrollerBase extends Component {
  * * `'hidden'`.
  *
  * @name horizontalScrollbar
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {String}
  * @default 'auto'
  * @public
@@ -271,7 +272,7 @@ class ScrollerBase extends Component {
  * Prevents scroll by wheeling on the scroller.
  *
  * @name noScrollByWheel
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {Boolean}
  * @default false
  * @public
@@ -285,7 +286,7 @@ class ScrollerBase extends Component {
  * Use `onScrollStart` or `onScrollStop` instead.
  *
  * @name onScroll
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {Function}
  * @param {Object} event
  * @param {Number} event.scrollLeft Scroll left value.
@@ -315,7 +316,7 @@ class ScrollerBase extends Component {
  * ```
  *
  * @name onScrollStart
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {Function}
  * @param {Object} event
  * @param {Number} event.scrollLeft Scroll left value.
@@ -345,7 +346,7 @@ class ScrollerBase extends Component {
  * ```
  *
  * @name onScrollStop
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {Function}
  * @param {Object} event
  * @param {Number} event.scrollLeft Scroll left value.
@@ -363,7 +364,7 @@ class ScrollerBase extends Component {
  * * `'hidden'`.
  *
  * @name verticalScrollbar
- * @memberof ui/Scroller.ScrollerBase.prototype
+ * @memberof ui/Scroller.ScrollerBasic.prototype
  * @type {String}
  * @default 'auto'
  * @public
@@ -379,7 +380,7 @@ class ScrollerBase extends Component {
  *
  * @class Scroller
  * @memberof ui/Scroller
- * @extends ui/Scroller.ScrollerBase
+ * @extends ui/Scroller.ScrollerBasic
  * @ui
  * @public
  */
@@ -407,7 +408,7 @@ const Scroller = (props) => {
 			<div {...scrollContainerProps}>
 				<div {...innerScrollContainerProps}>
 					<ChildWrapper {...childWrapperProps}>
-						<ScrollerBase {...childProps} />
+						<ScrollerBasic {...childProps} />
 					</ChildWrapper>
 					{isVerticalScrollbarVisible ? <Scrollbar {...verticalScrollbarProps} /> : null}
 				</div>
@@ -458,5 +459,6 @@ Scroller.defaultProps = {
 export default Scroller;
 export {
 	Scroller,
-	UiScrollerBase as ScrollerBase // to support legacy ScrollerBase
+	ScrollerBase, // to support legacy ScrollerBase
+	ScrollerBasic // to support theme libraries
 };
