+++ packages/ui/VirtualList/VirtualList.js
@@ -6,6 +6,7 @@
  * @exports VirtualGridList
  * @exports VirtualList
  * @exports VirtualListBase
+ * @exports VirtualListBasic
  */
 
 import PropTypes from 'prop-types';
@@ -16,14 +17,14 @@ import useScroll from '../useScroll';
 import Scrollbar from '../useScroll/Scrollbar';
 
 import {UiVirtualListBase, UiVirtualListBaseNative} from './UiVirtualListBase';
-import {gridListItemSizeShape, itemSizesShape, ThemeVirtualListBase as VirtualListBase} from './VirtualListBase';
+import {gridListItemSizeShape, itemSizesShape, VirtualListBasic} from './VirtualListBasic';
 
 /**
  * An unstyled scrollable virtual list component with touch support.
  *
  * @class VirtualList
  * @memberof ui/VirtualList
- * @extends ui/VirtualList.VirtualListBase
+ * @extends ui/VirtualList.VirtualListBasic
  * @ui
  * @public
  */
@@ -51,7 +52,7 @@ const VirtualList = ({role, ...rest}) => {
 			<div {...scrollContainerProps}>
 				<div {...innerScrollContainerProps}>
 					<ChildWrapper {...childWrapperProps}>
-						<VirtualListBase
+						<VirtualListBasic
 							{...childProps}
 							itemsRenderer={({cc}) => ( // eslint-disable-line react/jsx-no-bind
 								cc.length ? <div role={role}>{cc}</div> : null
@@ -131,7 +132,7 @@ VirtualList.defaultProps = {
  *
  * @class VirtualGridList
  * @memberof ui/VirtualList
- * @extends ui/VirtualList.VirtualListBase
+ * @extends ui/VirtualList.VirtualListBasic
  * @ui
  * @public
  */
@@ -155,7 +156,7 @@ const VirtualGridList = ({role, ...rest}) => {
 			<div {...scrollContainerProps}>
 				<div {...innerScrollContainerProps}>
 					<ChildWrapper {...childWrapperProps}>
-						<VirtualListBase
+						<VirtualListBasic
 							{...childProps}
 							itemsRenderer={({cc}) => ( // eslint-disable-line react/jsx-no-bind
 								cc.length ? <div role={role}>{cc}</div> : null
@@ -240,6 +241,7 @@ export {
 	itemSizesShape,
 	VirtualGridList,
 	VirtualList,
+	VirtualListBasic,
 	UiVirtualListBase as VirtualListBase, // to support legacy VirtualListBase
 	UiVirtualListBaseNative as VirtualListBaseNative // to support legacy VirtualListBaseNative
 };
